wlcore: change u32 to unsigned long in struct wl1271
Change u32 to unsigned long in struct wl1271 to make fields reg_ch_conf_last and
reg_ch_conf_pending 64bit aligned.

Without this, on 64bit ARM, wlcore_set_pending_regdomain_ch() will fail at
the point it calls set_bit(ch_bit_idx, (long*)wl->reg_ch_conf_pending);
Here is the error message while doing iw wlan0 scan or connect:

[  760.644782] wlcore: WL1271_ACX_INTR_DATA
[  760.648734] Unhandled fault: alignment fault (0x96000021) at 0xffffffc01bdc1b54
[  760.656048] Internal error: : 96000021 [#1] PREEMPT SMP
[  760.661267] Modules linked in: arc4 wl18xx wlcore mac80211 cfg80211 wlcore_sdio ti_conf_wl18xx btwilink st_drv bluetooth rfkill
[  760.672824] CPU: 0 PID: 1514 Comm: irq/198-wl18xx Tainted: G        W    3.10.52-wl-00204-gf704232-dirty #8
[  760.682558] task: ffffffc01d956040 ti: ffffffc01bea4000 task.ti: ffffffc01bea4000
[  760.690043] PC is at set_bit+0x14/0x30
[  760.693806] LR is at wlcore_set_pending_regdomain_ch+0x38/0x44 [wlcore]

+++ drivers/net/wireless/ti/wlcore/wlcore.h
@@ -320,10 +320,14 @@ struct wl1271 {
 	struct work_struct recovery_work;
 	bool watchdog_recovery;
 
+	/* change u32 to unsigned long to the field 64bit alignment
+	 * see commit message to learn more
+	 */
+
 	/* Reg domain last configuration */
-	u32 reg_ch_conf_last[2];
+	unsigned long reg_ch_conf_last[2];
 	/* Reg domain pending configuration */
-	u32 reg_ch_conf_pending[2];
+	unsigned long reg_ch_conf_pending[2];
 
 	/* Pointer that holds DMA-friendly block for the mailbox */
 	void *mbox;
